% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radafruitio.R
\docType{package}
\name{radafruitio}
\alias{radafruitio}
\alias{radafruitio-package}
\title{\href{https://io.adafruit.com}{\strong{Adafruit IO}} is a platform designed by
\href{https://www.adafruit.com}{Adafruit} to display, respond, and interact with
any internet of things (IoT) project's data. The
\href{https://io.adafruit.com}{\strong{Adafruit IO}} HTTP API provides access to
\href{https://io.adafruit.com}{\strong{Adafruit IO}} data from any programming language
or hardware environment that can speak HTTP. The \code{radafruitio} package
provides functions to interface with the
\href{https://io.adafruit.com}{\strong{Adafruit IO}} HTTP API through R.}
\description{
Adafruit IO is a platform designed by Adafruit to display, respond, 
    and interact with any internet of things (IoT) project's data. The Adafruit 
    IO HTTP API provides access to Adafruit IO data from any programming 
    language or hardware environment that can speak HTTP. The radafruitio
    package provides functions to interface with the Adafruit IO HTTP API
    through R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ernestguevarra/radafruitio}
  \item Report bugs at \url{https://github.com/ernestguevarra/radafruitio/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
